Rewrite docs table to TypeScript and remove unneeded parts for mdx color tables
